+++ web/src/ts/fallback_presenter/fallback_slide.ts
@@ -1,0 +1,8 @@
+import { Slide } from "../presenter/slide";
+
+export class FallbackSlide extends Slide {
+    override get_src_url(): string {
+        return this.video;
+    }
+}
+
