--- web/src/ts/video_viewer.ts
@@ -1,1 +1,0 @@
-import "../video_viewer.css"
