--- web/src/video_viewer.css
@@ -1,20 +1,0 @@
-body {
-    margin: 0px;
-    background-color: black;
-    width: 100vw;
-    height: 100vh;
-    display: flex;
-}
-
-div.playback {
-    display: flex;
-    min-width: 100vw;
-    max-width: 100vw;
-    width: 100vw;
-}
-
-video {
-    min-width: inherit;
-    max-width: inherit;
-    width: inherit;
-}
