--- web/src/video_viewer.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, height=device-height, initial-scale=1.0">
-    <title>Manim Web Presenter Controls</title>
-</head>
-
-<body>
-    <div class="playback">
-        <video controls>
-            <source src="2201830969_222290472_3962032669.mp4" type="video/mp4">
-            Your browser doesn't support HTML5 videos.
-        </video>
-    </div>
-</body>
-
-</html>
